series: handle number with floating value
For example, Jackass has movies labeled as 2.5 and 3.5.
Also handle variable length on the number of digits.

+++ components/list/li_templ.go
@@ -21,7 +21,7 @@ const (
 type LiProps struct {
 	Class         string
 	Items         int
-	Number        int
+	Number        float64
 	NumberColumns int
 	Style         ListStyle
 }
